+++ .gitignore
@@ -1,2 +1,8 @@
-/node_modules/
+# package directories
+node_modules
+jspm_packages
+
+# Serverless directories
+.serverless
+
 .env.json
