+++ package.json
@@ -13,6 +13,7 @@
     "body-parser": "^1.18.3",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
-    "node-fetch": "^2.1.2"
+    "node-fetch": "^2.1.2",
+    "query-string": "^6.1.0"
   }
 }
