+++ test/test-data.json
@@ -1,0 +1,56 @@
+{
+  "resource": "/rating",
+  "path": "/rating",
+  "httpMethod": "POST",
+  "headers": {
+    "Accept": "application/json,*/*",
+    "Accept-Encoding": "gzip,deflate",
+    "CloudFront-Forwarded-Proto": "https",
+    "CloudFront-Is-Desktop-Viewer": "true",
+    "CloudFront-Is-Mobile-Viewer": "false",
+    "CloudFront-Is-SmartTV-Viewer": "false",
+    "CloudFront-Is-Tablet-Viewer": "false",
+    "CloudFront-Viewer-Country": "US",
+    "Content-Type": "application/x-www-form-urlencoded",
+    "Host": "a3a09qrn74.execute-api.us-east-2.amazonaws.com",
+    "User-Agent": "Slackbot 1.0 (+https://api.slack.com/robots)",
+    "Via": "1.1 a9ced60f02a91a154a8631077a254a91.cloudfront.net (CloudFront)",
+    "X-Amz-Cf-Id": "JtykwS6RZWyn4J8JNQiswyp3-npcu15Li1A_QGmpz9zri_8hjESKVA==",
+    "X-Amzn-Trace-Id": "Root=1-5b22f92d-62a8e81966ca33fbfc614a0a",
+    "X-Forwarded-For": "54.164.194.74, 54.239.145.57",
+    "X-Forwarded-Port": "443",
+    "X-Forwarded-Proto": "https"
+  },
+  "queryStringParameters": null,
+  "pathParameters": null,
+  "stageVariables": null,
+  "requestContext": {
+    "resourceId": "b85ez3",
+    "resourcePath": "/rating",
+    "httpMethod": "POST",
+    "extendedRequestId": "IfvfLE_2CYcF0Mw=",
+    "requestTime": "14/Jun/2018:23:24:29 +0000",
+    "path": "/dev/rating",
+    "accountId": "786856905369",
+    "protocol": "HTTP/1.1",
+    "stage": "dev",
+    "requestTimeEpoch": 1529018669900,
+    "requestId": "163f622b-702a-11e8-a7ac-41f03d077086",
+    "identity": {
+      "cognitoIdentityPoolId": null,
+      "accountId": null,
+      "cognitoIdentityId": null,
+      "caller": null,
+      "sourceIp": "54.164.194.74",
+      "accessKey": null,
+      "cognitoAuthenticationType": null,
+      "cognitoAuthenticationProvider": null,
+      "userArn": null,
+      "userAgent": "Slackbot 1.0 (+https://api.slack.com/robots)",
+      "user": null
+    },
+    "apiId": "asdf"
+  },
+  "body": "token=0QlLvJwCbCXLdubiOsPt8qMH&team_id=T8UCGGR6C&team_domain=salt-and-fire&channel_id=C8V5P3C9K&channel_name=usa&user_id=U8V5S0KQV&user_name=garretthogan18&command=%2Frating&text=star+wars&response_url=https%3A%2F%2Fhooks.slack.com%2Fcommands%2FT8UCGGR6C%2F381799792500%2FL6kG6l2AOmWXgbsQVJoyGM7S&trigger_id=381924152867.300424569216.dd2ba8b343169921fc3487f0a30b8ecb",
+  "isBase64Encoded": false
+}
